unpacked-zips/railway_qr_server_notiononly/requirements.txt
fastapi
uvicorn[standard]
jinja2
python-dotenv
requests
stripe
qrcode
Pillow
python-slugify
notion-client
